+++ MouseLock/Configuration/ReleaseModifierTapBehavior.cs
@@ -1,0 +1,8 @@
+namespace MouseLock.Configuration;
+
+public enum ReleaseModifierTapBehavior
+{
+    None,
+    UntilWorldClick,
+    UntilNextTap,
+}